amplify-cli-core refactor cleanup (#1511)
* refactor: move rds input utils to amplify category api

* test: fix test imports

* chore: extract api

* chore: remove unused functions
