+++ packages/amplify-category-api/src/provider-utils/awscloudformation/service-walkthroughs/import-appsync-api-walkthrough.ts
@@ -7,7 +7,6 @@ import { getCfnApiArtifactHandler } from '../cfn-api-artifact-handler';
 import { serviceWalkthroughResultToAddApiRequest } from '../utils/service-walkthrough-result-to-add-api-request';
 import { writeSchemaFile } from '../utils/graphql-schema-utils';
 import {
-  constructDefaultGlobalAmplifyInput,
   ImportAppSyncAPIInputs,
   ImportedDataSourceType,
   ImportedRDSType,
@@ -17,6 +16,7 @@ import { PREVIEW_BANNER, category } from '../../../category-constants';
 import { storeConnectionSecrets, testDatabaseConnection, getSecretsKey } from '../utils/rds-secrets/database-secrets';
 import * as path from 'path';
 import { RDS_SCHEMA_FILE_NAME } from '@aws-amplify/graphql-transformer-core';
+import { constructDefaultGlobalAmplifyInput } from '../utils/rds-input-utils';
 import { getAPIResourceDir } from '../utils/amplify-meta-utils';
 import * as fs from 'fs-extra';
