+++ packages/amplify-graphql-transformer-core/src/utils/index.ts
@@ -5,11 +5,5 @@ export { getTable, getKeySchema, getSortKeyFieldNames } from './schema-utils';
 export { DEFAULT_SCHEMA_DEFINITION } from './defaultSchema';
 export { IAM_AUTH_ROLE_PARAMETER, IAM_UNAUTH_ROLE_PARAMETER } from './authType';
 export { getParameterStoreSecretPath } from './rds-secret-utils';
-export {
-  constructDefaultGlobalAmplifyInput,
-  constructRDSGlobalAmplifyInput,
-  getRDSDBConfigFromAmplifyInput,
-  readRDSGlobalAmplifyInput,
-} from './rds-input-utils';
 export { JSONUtilities } from './jsonUtilities';
 export const APICategory = 'api';
